new: #40, fix: #41, and some cleanup
